+++ CHANGES.rst
@@ -1,6 +1,11 @@
 Changes
 =======
 
+Next Release
+------------------
+
+* Suggest ``assertIsNone(a)`` for ``assertTrue(a is None)``, etc.
+
 0.9.0 (2018-05-14)
 ------------------